Move ActiveMarker logic into ActiveMarker
This component should render regardless of if a marker has been selected
or not - allowing us to give it more consistent styles in future.

This commit changes the props required by the component to allow it to
internally determine the data required to render the selected, previous
and next keyframe styles.