Create a context to manage playback time
There are many places in this app that will be concerned with the
current play state or current time of the master animation, as well as
being able to pause, resume or skip to a specific time in the animation.

Rather than set these all up as functions in the App component and drill
them to the components that require them using props, this commit
creates a React context to hold this functionality and a hook to
retrieve it where it's needed.

This is using a technique by Kent C Dodds:
https://kentcdodds.com/blog/how-to-use-react-context-effectively