Add a React Context to manage animations
There are at least two places in this App that will use animations that
are synchronised - the Preview component showing the user's animation
and the Playhead, which will show the animation's progress along the
Tracks.

Rather than add more functions to the App component to manage these
animations, this commit creates a React Context for this purpose.

Instead of drilling props down to the components that need animation,
they are now able to call a custom Hook to retrieve the functionality
they require, following a technique outlined by Kent C Dodds[1].

The functionality in this context uses the Web Animation API[2].
Components wanting to add a synchronised animation will retrieve the
addAnimation function from this context and call it, passing a DOM
element that will be animated and an array of Keyframes[3] describing
the styles to be animated.

It also includes pause and resume methods, that will perform those
operations on all Web Animations running in the document (so we should
avoid using web animations to animate anything that shouldn't be
synchronised with the Preview)

[1] https://kentcdodds.com/blog/how-to-use-react-context-effectively
[2] https://developer.mozilla.org/en-US/docs/Web/API/Web_Animations_API/Using_the_Web_Animations_API
[3] https://developer.mozilla.org/en-US/docs/Web/API/Web_Animations_API/Keyframe_Formats