Add allow-same-origin to iFrame sandbox
I've been around the houses trying to figure this one out.

By default, you cannot execute javascript from a parent window on
content inside an iFrame UNLESS the parent document and the document
inside the iframe are on the same domain (/origin).

Our iFrame content here is not on any domain, it's generated at runtime,
but I was still unable to access it with javascript and was shown a
crossorigin error.

The 'sandbox' property on the iFrame is a whitelist informing the
browser what functionality should be allowed inside the iFrame.  We want
to keep this to a minimum, as we're allowing users to enter their own
code, so we had this whitelist empty, meaning only minimal, basic
functionality was allowed.

Adding alow-same-origin to this whitelist allows the iFramed document to
retain the same origin as the parent window (our app), so we can run
javascript on it.

 > Why do we even need to do this?

Currently, the animations inside the Preview are run with CSS.  In order
to synchronise them with the Playhead, and pause/resume them, we need to
change them to use the Web Animations API - for this we need to be able
to access the elements using javascript.