Switch Preview to use Web Animations
In order to move away from CSS animations and use the AnimationManager
for animations in the Preview, I have had to:

Create a React ref
When the iframe is mounted, the ref now holds a reference to the DOM
node of this iframe.

Add an onload function
When the iframe has loaded it's content, we can use the created ref to
access the child document.

Create animations
Inside the onload function, we run querySelectors to get each element
targeted by each Track.  We also convert the TrackKeyframes to
web animation keyframes.  Where a CSS animation keyframe might look like
this:

  0%, 50% {
    opacity: 1;
  }
  25%, 75% {
    opacity: 0;
  }

Web Animations must have a single entry per percentage, with the
percentage value denoted by an 'offset' property and entries sorted
based on the offset. So the equivalent Web Animations keyframes array
looks like:

[
  { opacity: 1, offset: 0 },
  { opacity: 0, offset: 0.25 },
  { opacity: 1, offset: 0.5 },
  { opacity: 0, offset: 0.75 },
]

The Preview component now takes care of creating these keyframes from
the TrackKeyframes and passing them to the AnimationManager.