Store animations in AnimationManager
In order to pause, resume or otherwise modify all animations at the same
time, we need to keep track of all the animations we're creating.

This adds a React ref to hold references to each created animation.

When creating a new animation, any existing animations on the requested
element is cancelled and removed from the ref.

We can then run functionality on each of these animations at the same
time simply by iterating over the references in the ref.

Note that we're using a ref rather than state here because with state,
every update causes a component re-render.  These animations are
generally created _because of_ a render, so re-rendering would end us in
an infinite loop.

No components outside the AnimationManager need to access these
references, so we don't need a re-render. Using a ref to hold these
values means we don't get re-renders and the values persist.