Add playState to AnimationManager values
In order to track if the animation is paused or playing, we need the
AnimationManager to track the state of the animation.

This means overcoming the previous issue caused when using React state
inside the context, of re-renders being triggered by state changes.

This commit includes a workaround to the rerender issue.  When an
animation is created, we set it's currentTime to the same as the first
tracked animation.  If a component is rerendered and it's animation
recreated, we update its currentTime to put it in sync with the other
animations, meaning the users will not notice the rerender.

When animations are paused or resumed, we update the playState.

This commit also includes an exported enum which will allow consumers of
the context to determine the current playstate.