Enable ActiveMarker component to select new marker
The ActiveMarker component shows the CSS properties for the currently
active marker as well as the markers before and after the selected one.

This commit updates the ActiveMarker so that a user can click the next
or previous marker's contents to activate that marker in order to edit
it's contents (when we add marker content editing).